requirements.txt
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
mysql-connector-python==8.0.30
